requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.1
attrs==22.1.0
colorama==0.4.5
comtypes==1.1.11
iniconfig==1.1.1
MouseInfo==0.1.3
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.1.2
pytweening==1.0.4
pywin32==304
pywinauto==0.6.8
six==1.16.0
tomli==2.0.1
